+++ R/rtf_strwidth.R
@@ -89,10 +89,6 @@ rtf_strwidth <- function(tbl){
 
   db_list <- split(db, db$index)
   db_list <- lapply(db_list, function(x){
-                # family  <- x$family[1]
-                # if(grepl("Times New Roman", family)){
-                #   family = "Times"
-                # }
                 x$width <- graphics::strwidth(x$text, units = "inches", cex = x$cex[1], font = x$font[1], family = x$family[1])
                 x
   })